pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "djangosdk"
version = "0.1.0"
description = "Django-native AI SDK with agent-centric design, multi-provider support, and DRF integration"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Yusuf Ziya Cıvan" }]
keywords = ["django", "ai", "llm", "agents", "litellm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "django>=4.2",
    "litellm==1.82.6",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/ziyacivan/django-ai-sdk"
Repository = "https://github.com/ziyacivan/django-ai-sdk"
Issues = "https://github.com/ziyacivan/django-ai-sdk/issues"
Changelog = "https://github.com/ziyacivan/django-ai-sdk/blob/master/CHANGELOG.md"

[project.optional-dependencies]
drf = ["djangorestframework>=3.14"]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-asyncio>=0.23",
    "coverage>=7.0",
    "djangorestframework>=3.14",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["djangosdk"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.hatch.build.targets.wheel]
packages = ["djangosdk"]
